requirements.txt
Flask==1.1.1
Flask-Caching==1.8.0
marshmallow==3.3.0
pytest==5.3.3
python-dotenv==0.10.4
requests==2.22.0
responses==0.10.9
waitress==1.4.2